+++ src/Motor.php
@@ -7,5 +7,5 @@ class Motor
     /**
      * Versión del núcleo del motor.
      */
-    public const VERSION = '0.5.24';
+    public const VERSION = '0.5.25';
 }